+++ src/lib.rs
@@ -156,6 +156,8 @@ pub mod error;
 pub use error::{Error, InsertError, Result};
 
 pub mod io;
+mod list;
 mod loop_logic;
 mod macros;
 mod sources;
+mod token;